requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# AI API client
openai>=1.12.0

# Image processing
Pillow>=10.2.0

# Clipboard
pyperclip>=1.8.2

# Global hotkey
pynput>=1.7.6

# Data validation
pydantic>=2.5.0
